Refactors for external classes
