+++ plugintemplate.php
@@ -3,7 +3,7 @@
  * Plugin Name: Plugin Template
  * Plugin URI: https://github.com/MITLibraries/wp-plugin-template
  * Description: This is a sample plugin
- * Version: 1.0-beta7
+ * Version: 1.1
  * Author: Matt Bernhardt
  * Author URI: https://github.com/matt-bernhardt
  * License: GPL2
@@ -28,73 +28,22 @@
  * along with {Plugin Name}. If not, see {URI to Plugin License}.
  */
 
+namespace mitlib;
+
 // Don't call the file directly!
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-/**
- * Defines base widget.
- */
-class Plugin_Template extends WP_Widget {
-
-	/**
-	 * Constructor
-	 */
-	function __construct() {
-		$widget_ops = array(
-			'classname' => 'plugin-template-widget',
-			'description' => __( 'Template class for plugin', 'plugintemplate' ),
-		);
-		parent::__construct( 'plugin-template', __( 'Base Widget Template', 'plugintemplate' ), $widget_ops );
-	}
-
-	/**
-	 * Widget() builds the output
-	 *
-	 * @param array $args See WP_Widget in Developer documentation.
-	 * @param array $instance See WP_Widget in Developer documentation.
-	 * @link https://developer.wordpress.org/reference/classes/wp_widget/
-	 */
-	function widget( $args, $instance ) {
-		$args = null;
-		$instance = null;
-		echo '<p>This is my widget.</p>';
-	}
-}
+// Include the necessary classes.
+include_once( 'class-plugin-template.php' );
+include_once( 'class-plugin-template-extension.php' );
 
 /**
- * Registers base widget.
+ * Registers template widgets.
  */
 function register_template_widget() {
-	register_widget( 'Plugin_Template' );
+	register_widget( 'mitlib\Plugin_Template' );
+	register_widget( 'mitlib\Plugin_Template_Extension' );
 }
-add_action( 'widgets_init', 'register_template_widget' );
-
-/**
- * Stretch goal: This defines an extended widget, which overrides specific
- * methods of the base class. This is the heart of the "child plugin"
- * concept.
- */
-class Plugin_Template_Extension extends Plugin_Template {
-
-	/**
-	 * Widget() builds the output.
-	 *
-	 * @param array $args no idea.
-	 * @param array $instance no idea.
-	 */
-	function widget( $args, $instance ) {
-		$args = null;
-		$instance = null;
-		echo '<p>This is my <em>extended</em> widget.</p>';
-	}
-}
-
-/**
- * Registers extended widget.
- */
-function register_extended_widget() {
-	register_widget( 'Plugin_Template_Extension' );
-}
-add_action( 'widgets_init', 'register_extended_widget' );
+add_action( 'widgets_init', 'mitlib\register_template_widget' );
